refactor: shared BuyProductInteractive component across all buy pages
- Extract BuyProductInteractive into components/shop for reuse
- All buy pages (agrochemicals, plant-nutrition, animal-health, feeds) now use the same 3-column layout
- Extract AgrochemicalDosageTable component; plant-nutrition guide now auto-detects fertilizer vs agrochemical rate format (fixes biostimulants display)
- Remove old per-page ProductInteractive files
- Feed listing pagination scrolls to top on page change
