+++ apps/client-fnp/app/(landing)/buy-agrochemicals/[slug]/page.tsx
@@ -1,6 +1,8 @@
 import { queryAgroChemical } from "@/lib/query"
 import Link from "next/link"
-import { ProductInteractive } from "./ProductInteractive"
+import { Beaker } from "lucide-react"
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import { BuyProductInteractive } from "@/components/shop/BuyProductInteractive"
 
 interface BuyAgroChemicalPageProps {
     params: Promise<{ slug: string }>
@@ -46,16 +48,65 @@ export default async function BuyAgroChemicalPage({ params }: BuyAgroChemicalPag
         )
     }
 
+    const tabs = ["overview", "active-ingredients", ...(chemical.targets?.length > 0 ? ["targets"] : [])]
+
+    const tabsContent = (
+        <Tabs defaultValue="overview">
+            <TabsList className="w-full justify-start border-b rounded-none h-auto p-0 bg-transparent gap-1">
+                {tabs.map((tab) => (
+                    <TabsTrigger key={tab} value={tab} className="rounded-none border-b-2 border-transparent data-[state=active]:border-primary data-[state=active]:bg-transparent text-sm px-3 pb-2 capitalize">
+                        {tab.replace("-", " ")}
+                    </TabsTrigger>
+                ))}
+            </TabsList>
+            <TabsContent value="overview" className="mt-4">
+                <p className="text-sm text-muted-foreground leading-relaxed">
+                    {chemical.product_overview
+                        ? chemical.product_overview
+                        : chemical.agrochemical_category?.name
+                        ? <>This is a <span className="font-medium capitalize">{chemical.agrochemical_category.name}</span> designed for effective pest and disease control when used as directed.</>
+                        : `${chemical.name} is a professional agrochemical for crop protection.`}
+                </p>
+            </TabsContent>
+            <TabsContent value="active-ingredients" className="mt-4">
+                {chemical.active_ingredients?.length > 0 ? (
+                    <div className="divide-y">
+                        {chemical.active_ingredients.map((ai: any, idx: number) => (
+                            <div key={idx} className="flex justify-between py-2 text-sm">
+                                <span className="capitalize">{ai.name}</span>
+                                <span className="font-semibold">{ai.dosage_value} {ai.dosage_unit}</span>
+                            </div>
+                        ))}
+                    </div>
+                ) : (
+                    <p className="text-sm text-muted-foreground mt-4">No information available.</p>
+                )}
+            </TabsContent>
+            {chemical.targets?.length > 0 && (
+                <TabsContent value="targets" className="mt-4">
+                    <ul className="space-y-1.5">
+                        {chemical.targets.map((target: any, idx: number) => (
+                            <li key={idx} className="flex items-start gap-2 text-sm">
+                                <span className="h-1.5 w-1.5 rounded-full bg-green-500 flex-shrink-0 mt-1.5" />
+                                <span>
+                                    <span className="font-medium">{target.name}</span>
+                                    {target.scientific_name && (
+                                        <span className="text-xs text-muted-foreground italic ml-1">({target.scientific_name})</span>
+                                    )}
+                                </span>
+                            </li>
+                        ))}
+                    </ul>
+                </TabsContent>
+            )}
+        </Tabs>
+    )
+
     return (
         <div className="min-h-screen bg-background">
             {structuredData && (
-                <script
-                    type="application/ld+json"
-                    dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
-                />
+                <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }} />
             )}
-
-            {/* Breadcrumb */}
             <div className="border-b">
                 <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-3">
                     <nav className="flex text-sm text-muted-foreground">
@@ -67,9 +118,20 @@ export default async function BuyAgroChemicalPage({ params }: BuyAgroChemicalPag
                     </nav>
                 </div>
             </div>
-
             <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-                <ProductInteractive chemical={chemical} slug={slug} baseUrl={baseUrl} />
+                <BuyProductInteractive
+                    product={chemical}
+                    slug={slug}
+                    baseUrl={baseUrl}
+                    productType="agrochemical"
+                    categoryName={chemical.agrochemical_category?.name}
+                    brandHref={chemical.brand ? `/buy-agrochemicals?brand=${chemical.brand.id}` : undefined}
+                    shopHref="/buy-agrochemicals"
+                    guideHref={`/agrochemical-guides/${chemical.agrochemical_category?.slug || ""}/${slug}?from=buy-agrochemicals/${slug}`}
+                    loginRedirect={`/buy-agrochemicals/${slug}`}
+                    fallbackIcon={<Beaker className="w-28 h-28 text-muted-foreground/20" />}
+                    tabsContent={tabsContent}
+                />
             </div>
         </div>
     )
